requirements/all.txt
numpy
matplotlib
pandas
yfinance
feedparser
python-dotenv
statsmodels
scikit-learn
arch
openai